+++ tests/main/clients/custom_files_names/expected_client/base_model.py
@@ -7,6 +7,14 @@ from pydantic.fields import ModelField
 from .scalars import SCALARS_PARSE_FUNCTIONS, SCALARS_SERIALIZE_FUNCTIONS
 
 
+class UnsetType:
+    def __bool__(self) -> bool:
+        return False
+
+
+UNSET = UnsetType()
+
+
 class BaseModel(PydanticBaseModel):
     class Config:
         allow_population_by_field_name = True